requirements.txt
python-dotenv
praw
faker
findspark
psycopg2-binary
pyspark
kafka-python